Implement game archiving and loading.
This implementation is a little messy as object definitions are modified
slightly as required by jackson's serielisation. Some serielisation may
not work as expected, especially those contain Map<> as the fields. This
implementation is likely to break. The improvement will be to work
around the issue with reference to jackson.